requirements.txt
svgpathtools
imgkit
beautifulsoup4
google-generativeai
requests
python-dotenv
transformers
torch 
torchvision 
torchaudio